Worked on Issue #6
Created a new lib file called auth.ts and imported that to ap/auth/route.ts in order to let upload to vercel.
