+++ package.json
@@ -21,7 +21,7 @@
     "bcrypt": "^5.1.1",
     "bootstrap": "^5.3.3",
     "multiselect-react-dropdown": "^2.0.25",
-    "next": "14.2.10",
+    "next": "^14.2.33",
     "next-auth": "^4.24.7",
     "react": "^18",
     "react-bootstrap": "^2.10.4",
